+++ src/DuckFunkCore.php
@@ -2,10 +2,22 @@
 
 namespace Torralbodavid\DuckFunkCore;
 
+use Illuminate\Support\Facades\Auth;
+use Torralbodavid\DuckFunkCore\Models\Arcturus\User;
+
 class DuckFunkCore
 {
     /*
      * Package version
      */
-    const PACKAGE_VERSION = '0.0.2';
+    const PACKAGE_VERSION = '0.0.3';
+
+    protected User $user;
+
+    public function __construct()
+    {
+        $this->user = User::find(Auth::id());
+    }
+
+    public function user(): User { return User::find(Auth::id()); }
 }