+++ src/DuckFunkCoreServiceProvider.php
@@ -17,32 +17,33 @@ class DuckFunkCoreServiceProvider extends ServiceProvider
 
         // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'duck-funk-core');
         $this->loadViewsFrom(__DIR__.'/resources/views/'.config('duck-funk.template'), 'duck-funk-core');
-        // $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadViewsFrom(__DIR__.'/resources/views/housekeeping', 'housekeeping');
+        $this->loadViewsFrom(__DIR__.'/resources/views/auth', 'auth');
+
+        $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
         $this->loadRoutesFrom(__DIR__.'/routes/duck-funk.php');
 
-        if ($this->app->runningInConsole()) {
-            $this->publishes([
+        $this->publishes([
                 __DIR__.'/../config/config.php' => config_path('duck-funk.php'),
             ], 'config');
 
-            // Publishing the views.
-            /*$this->publishes([
-                __DIR__.'/resources/views' => resource_path('views/vendor/duck-funk-core'),
-            ], 'views');*/
+        // Publishing the views.
+        /*$this->publishes([
+            __DIR__.'/resources/views' => resource_path('views/vendor/duck-funk-core'),
+        ], 'views');*/
 
-            // Publishing assets.
-            $this->publishes([
+        // Publishing assets.
+        $this->publishes([
                 __DIR__.'/resources/assets/images' => public_path('vendor/duck-funk-core/images'),
             ], 'images');
 
-            // Publishing the translation files.
+        // Publishing the translation files.
             /*$this->publishes([
                 __DIR__.'/../resources/lang' => resource_path('lang/vendor/duck-funk-core'),
             ], 'lang');*/
 
             // Registering package commands.
             // $this->commands([]);
-        }
     }
 
     /**