+++ src/Http/Controllers/Housekeeping/HousekeepingController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Torralbodavid\DuckFunkCore\Http\Controllers\Housekeeping;
+
+class HousekeepingController
+{
+    public function __invoke()
+    {
+        return view('housekeeping::dashboard');
+    }
+}