+++ src/Models/Arcturus/Permissions.php
@@ -8,4 +8,10 @@ class Permissions extends Model
 {
     protected $table = 'permissions';
     protected $guarded = [];
+    public $timestamps = false;
+
+    public function canReadHousekeeping()
+    {
+        return $this->duck_funk_housekeeping_read;
+    }
 }