+++ src/helpers.php
@@ -1,0 +1,6 @@
+<?php
+
+function core()
+{
+    return app('duck-funk-core');
+}