+++ tests/TestCase.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Torralbodavid\DuckFunkCore\Tests;
+
+use Torralbodavid\DuckFunkCore\DuckFunkCoreServiceProvider;
+
+abstract class TestCase extends \Orchestra\Testbench\TestCase
+{
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->loadMigrationsFrom(__DIR__.'/../tests/database/migrations');
+        $this->withFactories(__DIR__.'/../database/factories');
+    }
+
+    protected function getPackageProviders($app)
+    {
+        return DuckFunkCoreServiceProvider::class;
+    }
+
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.default', 'testbench');
+        $app['config']->set('database.connections.testbench', [
+            'driver'   => 'sqlite',
+            'database' => ':memory:',
+            'prefix'   => '',
+        ]);
+    }
+}